Revert "Add support for negative numbers"
This reverts commit 757b2c89992e466082ec92aef293f5d28c813ec9.

I thought this was an easy implement without having to handle unary
opperations, this is not the case. x -1 would be wrongly interpreted
as a syntax error instead of an opperation. This has to be done on the
level of the parser, which has the required context to know the
differenece between a subtraction and a negation of a number.